third_party/blink/web_tests/external/wpt/css/css-color/parsing/color-invalid-rgb.html

<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8">
<title>CSS Color Level 4: Parsing and serialization of colors using invalid RGB notation</title>
<link rel="help" href="https://drafts.csswg.org/css-color-4/#rgb-functions">
<link rel="help" href="https://drafts.csswg.org/css-color-4/#resolving-sRGB-values">
<link rel="help" href="https://drafts.csswg.org/css-color-4/#serializing-sRGB-values">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/support/parsing-testcommon.js"></script>
</head>
<body>
<script>
test_invalid_value("color", "rgb(none, none, none)");
test_invalid_value("color", "rgba(none, none, none, none)");
test_invalid_value("color", "rgb(128, 0, none)");
test_invalid_value("color", "rgb(255, 255, 255, none)");
</script>
</body>
</html>
