third_party/blink/web_tests/external/wpt/css/css-color/parsing/color-invalid-rgb-expected.txt

This is a testharness.js-based test.
[FAIL] e.style['color'] = "rgb(0 0, 0)" should not set the property value
  assert_equals: expected "" but got "rgb(0, 0, 0)"
Harness: the test ran to completion.
